pyproject.toml
[project]
name = "mlb-pitch-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mlb-statsapi>=1.9.0",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tweepy>=4.16.0",
    "xgboost>=3.1.3",
]
